+++ mockup/patterns/pickadate/pattern.js
@@ -100,7 +100,9 @@ define([
       separator: ' ',
       date: {
         selectYears: true,
-        selectMonths: true
+        selectMonths: true,
+        formatSubmit: 'yyyy-mm-dd',
+        format: 'yyyy-mm-dd'
       },
       time: {
       },
@@ -144,7 +146,6 @@ define([
             .insertAfter(self.$el);
 
       if (self.options.date !== false) {
-        self.options.date.formatSubmit = 'yyyy-mm-dd';
         self.$date = $('<input type="text"/>')
               .attr('placeholder', self.options.placeholderDate)
               .attr('data-value', dateValue)
@@ -162,7 +163,7 @@ define([
                     }
                   }
                   if (e.hasOwnProperty('clear')) {
-                    self.$el.removeAttr('value');
+                    self.$el.val('');
                     self.$date.attr('data-value', '');
                   }
                 }
@@ -196,7 +197,7 @@ define([
                     }
                   }
                   if (e.hasOwnProperty('clear')) {
-                    self.$el.removeAttr('value');
+                    self.$el.val('');
                     self.$time.attr('data-value', '');
                   }
                 }
@@ -277,7 +278,7 @@ define([
             dateValue = self.$date.data('pickadate').get('select'),
             formatDate = date.formats.toString;
         if (dateValue) {
-          value += formatDate.apply(date, ['yyyy-mm-dd', dateValue]);
+          value += formatDate.apply(date, [self.options.date.formatSubmit, dateValue]);
         }
       }
 
@@ -301,7 +302,7 @@ define([
         }
       }
 
-      self.$el.attr('value', value);
+      self.$el.val(value);
 
       self.emit('updated');
     }